backend/requirements.txt
flask>=3.0
authlib>=1.3
requests>=2.31
Pillow>=10.0
gunicorn>=21.0
python-dotenv>=1.0
flask-limiter>=3.0
redis>=5.0
pytest>=8.0
pytest-flask>=1.3
